rename: entities user 도메인 이름 변경 (userInfo -> user)
