+++ src/entities/user/model/userInfoStore.tsx
