+++ src/views/connect/ui/ConnectStep.tsx
@@ -4,7 +4,7 @@ import { coupleApi, useRefetchCoupleInfo } from '@/entities/couple'
 import { shareWithKakao } from '@/features/share'
 import { useToast } from '@/shared/lib'
 import { BaseButton, DateButton, ProgressBar, TextButton } from '@/shared/ui'
-import useUserInfoStore from '@/entities/userInfo/model/userInfoStore'
+import useUserInfoStore from '@/entities/user/model/userInfoStore'
 import { useMutation } from '@tanstack/react-query'
 import { AxiosError, isAxiosError } from 'axios'
 import { format } from 'date-fns'
